configure.in (AC_C_INLINE): Added.
gcc/:
        * configure.in (AC_C_INLINE): Added.
	* configure: Regenerate.

gcc/java:
	* Makefile.in (INTLLIBS): New.
	(LIBS): Add above.
	(DEPLIBS): Ditto.

gcc/f:
	* Makefile.in(INTLLIBS): New macro.
	(LIBS): Add INTLLIBS.
	(DEPLIBS): Likewise.

gcc/ch:
	* Makefile.in: Changed to be processed by configure.
	(INTLLIBS): New macro.
	(LIBS): Add INTLLIBS.
	(DEPLIBS): Likewise.
	* config-lang(outputs): Specify ch/Makefile.

gcc/po:
	* Makefile.in.in: Add --define switch for xgettext.

gcc/intl:
	* Makefile.in: Add -DIN_GCC to DEFS, add appropriate -I switches.
	* dcgettext.c: Define _GNU_SOURCE before any system header is
	included. Guard getcwd declaration with HAVE_DECL_GETCWD.
	* loadmsgcat.c (_nl_load_domain): Cast st.st_size in compaison to
	size_t.
	* localealias.c: Define _GNU_SOURCE before any system header is
	included. Don't use return value of memcpy because it could be a
	macro defined in terms of bcopy.
	(read_alias_file): Remove HAVE_MEMCPY guard and just don't use the
	return value of memcpy. Cast argument to fgets and strlen to char *.
	* explodename.c (_nl_explode_name): Cast codeset to unsigned char **
	in call to _nl_normalize_codeset.
	* l10nflist.c: Define _GNU_SOURCE before any system header is
	included.

From-SVN: r34423

+++ gcc/intl/loadmsgcat.c
@@ -91,7 +91,7 @@ _nl_load_domain (domain_file)
 
   /* We must know about the size of the file.  */
   if (fstat (fd, &st) != 0
-      || (size = (size_t) st.st_size) != st.st_size
+      || (size = (size_t) st.st_size) != (size_t) st.st_size
       || size < sizeof (struct mo_file_header))
     {
       /* Something went wrong.  */
